third_party/WebKit/LayoutTests/inspector/console/console-assert-expected.txt

CONSOLE ERROR: line 9: 1
CONSOLE ERROR: line 10: a
Tests that console.assert() will dump a message and stack trace with source URLs and line numbers.

console-assert.html:4 Assertion failed: 1b @ console-assert.html:4a @ console-assert.html:10
console-assert.html:5 Assertion failed: a bb @ console-assert.html:5a @ console-assert.html:10
